apps/agents/pyproject.toml
[project]
name = "kuratormind-agents"
version = "0.1.0"
description = "KuratorMind AI — Multi-agent forensic backend for Indonesian Kurators"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.12.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.25.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "sse-starlette>=2.2.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]
